requirements.txt
ultralytics
streamlit
Pillow
numpy
onnxruntime-gpu
scikit-image
imgaug
pyclipper
lmdb
tqdm
visualdl
rapidfuzz
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
cython
lxml
premailer
openpyxl
attrdict
Polygon3
lanms-neo==1.0.2
PyMuPDF<1.21.0